MAIN-78 :: postreview: remove theme, rename shared components && clean up
