--- config/storybook.js
@@ -1,8 +1,0 @@
-// import { configure } from '@storybook/react';
-//
-// function loadStories() {
-//   require('./../src/storybook/api-keys.story.tsx');
-//   // You can require as many stories as you need.
-// }
-//
-// configure(loadStories, module);
