+++ src/modules/api-keys/components/sidebar/sidebar.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { SFC } from 'react';
-import { Navigation } from '../../../shared-components/navigation';
-import { NavItem } from './theme/sidebar.css';
+import { Navigation } from '../../../components/navigation/navigation';
+import { NavItem } from './sidebar.css';
 
 const items = [
   {
