+++ src/modules/components/main-layout/__tests__/main-layout-test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Layout } from 'antd';
 import { shallow } from 'enzyme';
-import { Navigation } from '../../navigation';
+import { Navigation } from '../../navigation/navigation';
 import { MainLayout } from '../main-layout';
 const { Header, Sider, Content } = Layout;
