+++ src/pages/api-keys.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { SFC } from 'react';
 import { ApiKeysContainer } from '../modules/api-keys/containers/api-keys/api-keys.container';
-import { MainLayout } from '../modules/shared-components/main-layout/main-layout';
+import { MainLayout } from '../modules/components/main-layout/main-layout';
 import { SidebarNavigation } from '../modules/api-keys/components/sidebar/sidebar';
 
 export const ApiKeysLayout: SFC = () => (
